refactor(Components JS): Moved the component pages to their relevant directories
