+++ ui/src/main/js/global/components/audit/AuditPage.js
@@ -13,8 +13,8 @@ import {
 import AutoRefresh from 'component/common/AutoRefresh';
 import DescriptorLabel from 'component/common/DescriptorLabel';
 import RefreshTableCellFormatter from 'component/common/RefreshTableCellFormatter';
-import NotificationTypeLegend from 'dynamic/loaded/audit/NotificationTypeLegend';
-import AuditDetails from 'dynamic/loaded/audit/Details';
+import NotificationTypeLegend from 'global/components/audit/NotificationTypeLegend';
+import AuditDetails from 'global/components/audit/Details';
 import CheckboxInput from 'field/input/CheckboxInput';
 import * as DescriptorUtilities from 'util/descriptorUtilities';
 import ConfigurationLabel from 'component/common/ConfigurationLabel';
