+++ ui/src/main/js/global/components/audit/Details.js
