+++ ui/src/main/js/global/components/audit/NotificationTypeLegend.js
