+++ ui/src/main/js/global/components/user/UserManagement.js
@@ -3,8 +3,8 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { Tab, Tabs } from 'react-bootstrap';
 import ConfigurationLabel from 'component/common/ConfigurationLabel';
-import RoleTable from 'dynamic/loaded/users/RoleTable';
-import UserTable from 'dynamic/loaded/users/UserTable';
+import RoleTable from 'global/components/user/RoleTable';
+import UserTable from 'global/components/user/UserTable';
 import * as DescriptorUtilities from 'util/descriptorUtilities';
 import { USER_MANAGEMENT_INFO } from 'global/components/user/UserModel';
